% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_ds}
\alias{compare_ds}
\title{Compare data sets}
\usage{
compare_ds(ds1, ds2, precision = NULL, rounding = NULL)
}
\arguments{
\item{ds1}{The first dataset to compare.}

\item{ds2}{The second dataset to compare.}

\item{precision}{The desired number of decimal places to compare.}

\item{rounding}{The rounding function to be used in the comparison.}
}
\value{
The results of the comparison.
}
\description{
A function to compare two data sets  This function
may be used to compare R and SAS output.  It has a parameters to
control the desired precision and the rounding to be used.
}
