% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplements.R
\name{opts}
\alias{opts}
\title{Specifies options}
\usage{
opts(...)
}
\arguments{
\item{...}{Various options.}
}
\value{
An options class with the requested options.
}
\description{
The \code{output} function is a generic options collection.
Used on multiple functions and parameters.
}
\details{
Here is some details
}
