% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplements.R
\name{out}
\alias{out}
\title{Function to request an output dataset}
\usage{
out(stats = NULL, direction = "long", ...)
}
\arguments{
\item{stats}{The requested statistics. The available statistics depend
on the function the output parameter applies to.}

\item{direction}{How the output dataset should be organized.  Valid
values are "long" or "wide".  The default is "long".}

\item{...}{Various options.}
}
\value{
The output specifications.
}
\description{
The \code{out} function is used to request an output
dataset.  The function is used as a parameter to multiple procedures in
the \strong{procs} package.
}
\details{
Here is some details
}
