% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_print.R
\name{proc_print}
\alias{proc_print}
\title{Prints a dataset}
\usage{
proc_print(
  data,
  titles = NULL,
  report_type = "HTML",
  report_location = NULL,
  view = TRUE
)
}
\arguments{
\item{data}{The data to print.}

\item{titles}{A vector of titles.}

\item{report_type}{The type of report to create.}

\item{report_location}{The location of the report to create.}

\item{view}{Whether to send the print output to the viewer.  Valid
values are TRUE and FALSE.}
}
\description{
The \code{proc_print} function is used to print a dataset
or datasets.  To print multiple datasets, pass them to the \code{data}
parameter in a list.
By default, the function prints to the viewer. It may also be used
to print to the file system using the \code{report_type} and
\code{report_location} parameters.
}
\details{
Here is some details
}
