% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_reg.R
\encoding{UTF-8}
\name{proc_reg}
\alias{proc_reg}
\title{Calculates a Regression}
\usage{
proc_reg(
  data,
  model = NULL,
  by = NULL,
  output = NULL,
  options = NULL,
  titles = NULL
)
}
\arguments{
\item{data}{The input data frame for which to calculate summary statistics.
This parameter is required.}

\item{model}{A vector of paired variables to perform a paired T-Test on.
Variables should
be separated by a star (*).  The entire string should be quoted, for example,
\code{paired = "var1 * var2"}.  To test multiple pairs, place the pairs in a
quoted vector
: \code{paired = c("var1 * var2", "var3 * var4")}.  The parameter does not
accept parenthesis, hyphens, or any other shortcut syntax.}

\item{by}{An optional by group. If you specify a by group, the input
data will be subset on the by variable(s) prior to performing any
statistics.}

\item{output}{Whether or not to return datasets from the function. Valid
values are "out", "none", and "report".  Default is "out", and will
produce dataset output specifically designed for programmatic use. The "none"
option will return a NULL instead of a dataset or list of datasets.
The "report" keyword returns the datasets from the interactive report, which
may be different from the standard output. The output parameter also accepts
data shaping keywords "long, "stacked", and "wide".
These shaping keywords control the structure of the output data. See the
\strong{Data Shaping} section for additional details. Note that
multiple output keywords may be passed on a
character vector. For example,
to produce both a report dataset and a "long" output dataset,
use the parameter \code{output = c("report", "out", "long")}.}

\item{options}{A vector of optional keywords. Valid values are: "alpha =",
"h0 =", and "noprint". The "alpha = " option will set the alpha
value for confidence limit statistics.  The default is 95\% (alpha = 0.05).
The "h0 = " option sets the baseline hypothesis value for single-variable
hypothesis testing.  The "noprint" option turns off the interactive report.}

\item{titles}{A vector of one or more titles to use for the report output.}
}
\value{
Normally, the requested T-Test statistics are shown interactively
in the viewer, and output results are returned as a list of data frames.
You may then access individual datasets from the list using dollar sign ($)
syntax.
The interactive report can be turned off using the "noprint" option, and
the output datasets can be turned off using the "none" keyword on the
\code{output} parameter.
}
\description{
The \code{proc_reg} function performs a regression
for selected variables on the input dataset.
The variables are identified on the \code{var} parameter or the \code{paired}
parameter. The function will calculate a standard set of T-Test statistics.
Results are displayed in
the viewer interactively and returned from the function.
}
\details{
The \code{proc_reg} function is for ...
}
\section{Interactive Output}{

By default, \code{proc_ttest} results will
be sent to the viewer as an HTML report.  This functionality
makes it easy to get a quick analysis of your data. To turn off the
interactive report, pass the "noprint" keyword
to the \code{options} parameter.

The \code{titles} parameter allows you to set one or more titles for your
report.  Pass these titles as a vector of strings.

The exact datasets used for the interactive report can be returned as a list.
To return these datasets, pass
the "report" keyword on the \code{output} parameter. This list may in
turn be passed to \code{\link{proc_print}} to write the report to a file.
}

\section{Dataset Output}{

Dataset results are also returned from the function by default.
\code{proc_ttest} typically returns multiple datasets in a list. Each
dataset will be named according to the category of statistical
results.  There are three standard categories: "Statistics",
"ConfLimits", and "TTests". For the class style analysis, the function
also returns a dataset called "Equality" that shows the Folded F analysis.

The output datasets generated are optimized for data manipulation.
The column names have been standardized, and additional variables may
be present to help with data manipulation.  For example, the by variable
will always be named "BY".  In addition, data values in the
output datasets are intentionally not rounded or formatted
to give you the most accurate numeric results.
}

\section{Options}{

The \code{proc_reg} function recognizes the following options.  Options may
be passed as a quoted vector of strings, or an unquoted vector using the
\code{v()} function.
\itemize{
\item{\strong{alpha = }: The "alpha = " option will set the alpha
value for confidence limit statistics.  Set the alpha as a decimal value
between 0 and 1.  For example, you can set a 90\% confidence limit as
\code{alpha = 0.1}.
}
\item{\strong{h0}: The "h0 =" option is used to set the baseline mean value
for testing a single variable.  Pass the option as a name/value pair,
such as \code{h0 = 95}.
}
\item{\strong{noprint}: Whether to print the interactive report to the
viewer.  By default, the report is printed to the viewer. The "noprint"
option will inhibit printing.  You may inhibit printing globally by
setting the package print option to false:
\code{options("procs.print" = FALSE)}.
}
}
}

\section{Data Shaping}{

The output datasets produced by the function can be shaped
in different ways. These shaping options allow you to decide whether the
data should be returned long and skinny, or short and wide. The shaping
options can reduce the amount of data manipulation necessary to get the
data into the desired form. The
shaping options are as follows:
\itemize{
\item{\strong{long}: Transposes the output datasets
so that statistics are in rows and variables are in columns.
}
\item{\strong{stacked}: Requests that output datasets
be returned in "stacked" form, such that both statistics and
variables are in rows.
}
\item{\strong{wide}: Requests that output datasets
be returned in "wide" form, such that statistics are across the top in
columns, and variables are in rows. This shaping option is the default.
}
}
These shaping options are passed on the \code{output} parameter.  For example,
to return the data in "long" form, use \code{output = "long"}.
}

\examples{
# Turn off printing for CRAN checks
options("procs.print" = FALSE)

# Prepare sample data
dat1 <- subset(sleep, group == 1, c("ID", "extra"))
dat2 <- subset(sleep, group == 2, c("ID", "extra"))
dat <- data.frame(ID = dat1$ID, group1 = dat1$extra, group2 = dat2$extra)

# View sample data
dat
#    ID group1 group2
# 1   1    0.7    1.9
# 2   2   -1.6    0.8
# 3   3   -0.2    1.1
# 4   4   -1.2    0.1
# 5   5   -0.1   -0.1
# 6   6    3.4    4.4
# 7   7    3.7    5.5
# 8   8    0.8    1.6
# 9   9    0.0    4.6
# 10 10    2.0    3.4

}
