% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_transpose.R
\encoding{UTF-8}
\name{proc_transpose}
\alias{proc_transpose}
\title{Transposes a Data Frame}
\usage{
proc_transpose(
  data,
  by = NULL,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{by}{An optional by group.}

\item{var}{The variable or variables to transpose.}

\item{id}{The option specifications for the table parameter.}

\item{idlabel}{The variable(s) to calculate summary statistics for.}

\item{copy}{A vector of variable to retain in the output data frame
without transposition.}
}
\value{
A data frames that contains the transposed data.
}
\description{
Here is a sample function.
}
\details{
Some details about the sample function.
}
