% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestplots.R
\name{ttestplot}
\alias{ttestplot}
\title{Request T-Test Plots}
\usage{
ttestplot(type = "default", panel = TRUE, showh0 = FALSE)
}
\arguments{
\item{type}{The type(s) of plot to create. Multiple types should be passed
as a vector of strings.  Valid values are "agreement", "boxplot", "histogram",
"interval", "profiles", "qqplot", "summary".  The default value is
a vector with "summary" and "qqplot".}

\item{panel}{Whether or not to display the summary plot combined into
in a single panel.  Default is TRUE.  A value of FALSE will create
individual plots instead.  This parameter is equivalent to the
"unpack" keyword in SAS.}

\item{showh0}{Whether or not to show the h0 line on a single-variable T-test.
Default is FALSE.}
}
\description{
A function to request T-test plots on a call to
\code{\link{proc_ttest}}. The function allows you to specify the type of
regression plots to produce. It produces a combined diagnostics panel
by default.  You may also specify individual plots to create by setting
the "panel" parameter to FALSE, and passing a vector of plot names
on the "type" parameter.
}
\details{
Any requested
plots will be displayed on interactive reports only.
Plots are created as jpeg files, and stored in a temp directory.  Those
temporary files are then referenced by the interactive report to display
the graphic.

If desired, you may
output the report objects and pass to \code{\link{proc_print}}. To do this,
set \code{output = report} on the call to \code{\link{proc_freq}}, and pass
the entire list to \code{\link{proc_print}}.
}
\section{Plots}{

The \code{plots} parameter allows you to request several types of T-Test
plots. Below are the types of plots that are supported.  The list shows
the plot type keyword needed to request the plot, and a brief description:
\itemize{
\item{\strong{agreement}:  An agreement plot for the input data.  Only
available for paired comparisons.
}
\item{\strong{boxplot}: Displays a box and whisker plot for group comparisons,
including confidence intervals (if appropriate).
}
\item{\strong{histogram}: A histogram with normal curve and kernel density
overlays.
}
\item{\strong{interval}: Visualizes the confidence intervals for means.
}
\item{\strong{profiles}: A line plot mapping responses
between analysis variables. Available only for paired analysis.
}
\item{\strong{qqplot}: A quantile-quantile plot used to assess the assumption
of normality.
}
\item{\strong{summary}: Combines the histogram and boxplot charts
onto the same panel.
}
}
The above plots may be requested as a vector of keywords to the \code{plots}
parameter on \code{\link{proc_ttest}}, or as vector of values
to the \code{type} parameter on \code{ttestplots}.

Note that, when passed as a vector of keywords, only the requested plots
will be produced.  That is to say, the "only" keyword in SAS is implied
at all times for \code{\link{proc_ttest}}.
}

\examples{
library(procs)

# Turn off printing for CRAN checks
options("procs.print" = FALSE)

# View sample data
sleep
#    extra group ID
# 1    0.7     1  1
# 2   -1.6     1  2
# 3   -0.2     1  3
# 4   -1.2     1  4
# 5   -0.1     1  5
# 6    3.4     1  6
# 7    3.7     1  7
# 8    0.8     1  8
# 9    0.0     1  9
# 10   2.0     1 10
# 11   1.9     2  1
# 12   0.8     2  2
# 13   1.1     2  3
# 14   0.1     2  4
# 15  -0.1     2  5
# 16   4.4     2  6
# 17   5.5     2  7
# 18   1.6     2  8
# 19   4.6     2  9
# 20   3.4     2 10


# Example 1: plots = TRUE
proc_ttest(sleep,
           var = "extra",
           class = "group",
           plots = TRUE)

# Example 2: plots = "all"
proc_ttest(sleep,
           var = "extra",
           class = "group",
           plots = "all")

# Example 3: plots = {vector}
proc_ttest(sleep,
           var = "extra",
           class = "group",
           plots = c("histogram", "boxplot", "interval"))

# Example 4: plots = ttestplot()
proc_ttest(sleep,
           var = "extra",
           class = "group",
           plots = ttestplot("summary", panel = FALSE))

}
