% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{v}
\alias{v}
\title{Combine unquoted values}
\usage{
v(...)
}
\arguments{
\item{...}{One or more unquoted values.}
}
\value{
A vector of quoted values.
}
\description{
A function to combine unquoted values into a vector.  The
function will return a vector of quoted values.
}
\examples{
# Combine unquoted values
v(var1, var2, var3)
# [1] "var1" "var2" "var3"
}
