% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqplots.R
\name{freqplot}
\alias{freqplot}
\title{Request a Frequency Plot}
\usage{
freqplot(
  type = "barchart",
  orient = "vertical",
  scale = "freq",
  twoway = "groupvertical",
  groupby = "column",
  npanelpos = 4
)
}
\arguments{
\item{type}{The type of plot to create. Valid values are "barchart" or
"dotplot".  Default is "barchart".}

\item{orient}{The orientation of the plot.  Valid values are "vertical"
or "horizontal".  Default is "vertical".}

\item{scale}{The scale to use for the plot. Valid values are "freq",
"log", "percent", or "sqrt".  Default is "freq".  For two-way tables,
the value "grouppercent" is also valid.}

\item{twoway}{Options for two-way layouts. Valid values are
"cluster", "grouphorizontal", "groupvertical", or "stacked". Default
is "groupvertical".  This parameter applies to two-way tables only.}

\item{groupby}{The variable to group by. Valid values are "column" or "row".
Default is "column". This parameter applies to two-way tables only.}

\item{npanelpos}{The number of sections per panel. Default is 4. This
parameter applies to two-way tables only.}
}
\value{
The frequency plot object.
}
\description{
A function to request a frequency plot on a call to
\code{proc_freq}. The function allows you to specify the type of
frequency plot to produce, and various layout options. Any requested
plots will be displayed on interactive reports only. You may, however,
output the report objects and pass to \code{\link{proc_print}}.
}
