% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplements.R
\name{out}
\alias{out}
\title{Function to request an output dataset}
\usage{
out(
  stats = NULL,
  shape = NULL,
  report = FALSE,
  where = NULL,
  drop = NULL,
  keep = NULL,
  rename = NULL,
  format = NULL,
  label = NULL,
  table = NULL,
  ...
)
}
\arguments{
\item{stats}{The requested statistics. The available statistics depend
on the function the output parameter applies to.}

\item{shape}{How the output dataset should be organized.  Valid
values are "long", "wide", or "stacked".  The default is "long".}

\item{report}{Whether to output the tables produced for the
procedure report.  Valid values are TRUE and FALSE.  Default is FALSE.
If the parameter is TRUE, the function will output all datasets
produced by the report functionality, and in the same order.  If the
parameter is TRUE, all other output options will be ignored.}

\item{where}{An expression used to filter the output dataset.  Use
\code{expression} function to define the where clause.  The where
clause will be applied prior to other output options such as drop, keep,
or rename.}

\item{drop}{A vector of variables to drop from the output dataset.}

\item{keep}{A vector of variables to keep on the output dataset.
This parameter can also be used to order the output columns.}

\item{rename}{A named vector of columns to rename.  The name of the vector
item should correspond to the original output column name.  The value
of the vector should correspond to the new variable name.}

\item{format}{A named list of formats to assign to the output dataset.}

\item{label}{A named vector of labels to assign to the output dataset.}

\item{table}{A table request.  This parameter is used only for the
\code{\link{proc_freq}} function.}

\item{...}{Various options.  When used with \code{\link{proc_freq}},
the \code{table} option is most frequently used.}
}
\value{
The output specifications.
}
\description{
The \code{out} function is used to request an output
dataset.  The function is used as a parameter to multiple procedures in
the \strong{procs} package.
}
\details{
Here is some details
}
