% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\encoding{UTF-8}
\name{proc_freq}
\alias{proc_freq}
\title{A sample function}
\usage{
proc_freq(
  data,
  by = NULL,
  tables = NULL,
  table_options = NULL,
  test = NULL,
  test_options = NULL,
  weight = NULL,
  weight_options = NULL,
  print = "HTML",
  print_location = NULL,
  titles = NULL,
  digits = 4
)
}
\arguments{
\item{data}{The input data frame to perform frequency calculations on.}

\item{by}{An optional by group.}

\item{tables}{The variable or variables to perform frequency counts on.}

\item{table_options}{The option specifications for the table parameter.}

\item{test}{The test to use for the frequency.}

\item{test_options}{The option specifications for the test parameter.}

\item{weight}{An optional weight parameter.}

\item{weight_options}{The option specifications for the weight parameter.}

\item{print}{The output type for any report output.  Valid values are
'HTML', 'TXT', 'PDF', 'RTF', 'DOCX', or 'none'.  Multiple outputs can be
requested by passing a vector of output types.  If the print_location parameter
is specified, the outputs will be written to that location.  Otherwise,
they will be written to a temp directory.  The default value is 'HTML'.}

\item{print_location}{A full path to write any report output requested by
the print parameter. If the location is not specified, files will be
written to a temp directory.  Default is NULL.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{digits}{The number of digits to round statistics produced
by the function.}
}
\value{
A list of data frames that contain the requested frequency tables.
}
\description{
Here is a sample function.
}
\details{
Some details about the sample function.
}
