% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\encoding{UTF-8}
\name{proc_means}
\alias{proc_means}
\title{Calculates Summary Statistics}
\usage{
proc_means(
  data,
  var = NULL,
  stats = NULL,
  view = TRUE,
  report_type = NULL,
  report_location = NULL,
  report_style = NULL,
  titles = NULL,
  piped = FALSE
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{var}{The variable(s) to calculate summary statistics for.}

\item{stats}{The a quoted vector of summary statistics keywords.  Valid
keywords are: "n", "means", "median", "std", "min", "max", and "range".}

\item{view}{Whether to display the report in the interactive viewer.  Valid
values are TRUE and FALSE.  Default is TRUE.}

\item{report_type}{The output type for any report output.  Valid values are
'HTML', 'TXT', 'PDF', 'RTF', 'DOCX'.  Multiple outputs can be
requested by passing a vector of output types.  If the report_location parameter
is specified, the outputs will be written to that location.  Otherwise,
they will be written to a temp directory.  The default value is NULL.}

\item{report_location}{A path to write any report output requested by
the print parameter. The path may be either a full path with file name,
or a directory name.  If the file name is not specified, the procedure
type will be used.  If no path is specified, files will be
written to a temp directory.  Default is NULL.}

\item{report_style}{A theme name or style object to use on the report output.
See the \code{\link[reporter]{create_style}} in the \strong{reporter}
package for additional information on styles.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{piped}{Whether or not the \code{proc_freq} function is part of a data
pipeline.  Set this parameter to TRUE if you want the function to return
a single dataset instead of a list of datasets.  If there is more than one
table requested, the function will return the last requested table.}
}
\value{
A list of data frames that contain the requested frequency tables.
}
\description{
Here is a sample function.
}
\details{
Some details about the sample function.
}
