% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\encoding{UTF-8}
\name{proc_means}
\alias{proc_means}
\title{Calculates Summary Statistics}
\usage{
proc_means(
  data,
  by = NULL,
  class = NULL,
  class_options = NULL,
  var = NULL,
  weight = NULL,
  weight_options = NULL,
  view = TRUE,
  output = NULL,
  report_type = NULL,
  report_location = NULL,
  titles = NULL,
  digits = 4
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{by}{An optional by group.}

\item{class}{The variable or variables to perform frequency counts on.}

\item{class_options}{The option specifications for the table parameter.}

\item{var}{The variable(s) to calculate summary statistics for.}

\item{weight}{An optional weight parameter.}

\item{weight_options}{The option specifications for the weight parameter.}

\item{view}{Whether to display the report}

\item{output}{The directory path or libname to output data generated by
the \code{proc_freq} function.}

\item{report_type}{The output type for any report output.  Valid values are
'HTML', 'TXT', 'PDF', 'RTF', 'DOCX'.  Multiple outputs can be
requested by passing a vector of output types.  If the report_location parameter
is specified, the outputs will be written to that location.  Otherwise,
they will be written to a temp directory.  The default value is NULL.}

\item{report_location}{A path to write any report output requested by
the print parameter. The path may be either a full path with file name,
or a directory name.  If the file name is not specified, the procedure
type will be used.  If no path is specified, files will be
written to a temp directory.  Default is NULL.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{digits}{The number of digits to round statistics produced
by the function.}
}
\value{
A list of data frames that contain the requested frequency tables.
}
\description{
Here is a sample function.
}
\details{
Some details about the sample function.
}
