% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\encoding{UTF-8}
\name{proc_means}
\alias{proc_means}
\title{Calculates Summary Statistics}
\usage{
proc_means(
  data,
  by = NULL,
  class = NULL,
  var = NULL,
  stats = c("n", "mean", "std", "min", "max"),
  weight = NULL,
  view = TRUE,
  titles = NULL,
  options = NULL,
  ...,
  output = NULL
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{by}{An optional by group. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{class}{The \code{class} parameter is similar to the \code{by}
parameter, but the output is different.  By groups will create completely
separate tables, while class groups will be continued in the same table.
When a \code{by} and a \code{class} are both specified, the \code{class}
will be nested in the \code{by}.}

\item{var}{The variable(s) to calculate summary statistics for. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{stats}{A quoted vector of summary statistics keywords.  Valid
keywords are: "css", "clm", "cv", "n", "lclm", "mean", "median", "mode",
"min", "max",
"nmiss", "nobs", "range", "std", "stderr", "sum", "uclm", "var",
"q1", "q3", "p1", "p5", "p10", "p20", "p25", "p30", "p40",
"p50", "p60", "p70", "p75", "p80", "p90",
"p95", "p99", "qrange". You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function. See the \link{ProcMeansOptions} page for details on
the available options.}

\item{weight}{An optional weight parameter.}

\item{view}{Whether to display the report in the interactive viewer.  Valid
values are TRUE and FALSE.  Default is TRUE.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{options}{An options object created with the \code{\link{opts}}
function. This parameter can accept a variety of options.}

\item{...}{One or more output requests. To request a dataset output,
use the desired dataset name as the parameter name, and pass in the
\code{\link{out}} function with the desired specifics.  If only one
output dataset is requested, the function will return a single data frame.
If more than one output dataset is requested, the function will return
a list of data frames. If no statistics are specified in the \code{\link{out}}
function, the procedure will default to the statistics assigned on
the \code{stats} parameter.  Note that this default behavior is different
from SAS®.}

\item{output}{A named list of output data requests.  The name of the list item
will become the name of the item in the return list, if there is one.  The
value of the list item is a output request created with the \code{\link{out}}
function.  Output data requests may also be made on the \code{...} parameter
for convenience.  This parameter is provided in case you need to generate
output requests dynamically inside a function.}
}
\value{
The requested summary statistics.
}
\description{
The \code{proc_means} function generates summary statistics
for selected variables on the input dataset.  The variables are identified
on the \code{var} parameter.  The statistics to perform are identified
on the \code{stats} parameter.  By default, results are displayed in
the viewer interactively, and returned as a dataset or a list of datasets.
}
\details{
Some details about the sample function.

All statistics remove missing values by default.  Should there be an
option to include them?
}
\section{Options}{

\itemize{
\item{\strong{CSS}: Corrected Sum of Squares.

}
\item{\strong{CV}: Coefficient of Variation.

}
}
}

\section{Using Weights}{

Need a section explaining how to use weights.
}

\section{Data Constraints}{

Explain limits of data for each stat option.  Number of non-missing
values, etc.
}

