% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\encoding{UTF-8}
\name{proc_means}
\alias{proc_means}
\title{Calculates Summary Statistics}
\usage{
proc_means(
  data,
  by = NULL,
  class = NULL,
  var = NULL,
  stats = c("n", "mean", "std", "min", "max"),
  weight = NULL,
  titles = NULL,
  options = NULL
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{by}{An optional by group. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{class}{The \code{class} parameter is similar to the \code{by}
parameter, but the output is different.  By groups will create completely
separate tables, while class groups will be continued in the same table.
When a \code{by} and a \code{class} are both specified, the \code{class}
will be nested in the \code{by}.}

\item{var}{The variable(s) to calculate summary statistics for. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{stats}{A quoted vector of summary statistics keywords.  Valid
keywords are: "css", "clm", "cv", "n", "lclm", "mean", "median", "mode",
"min", "max",
"nmiss", "nobs", "range", "std", "stderr", "sum", "uclm", "var",
"q1", "q3", "p1", "p5", "p10", "p20", "p25", "p30", "p40",
"p50", "p60", "p70", "p75", "p80", "p90",
"p95", "p99", "qrange". You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function. See the \link{ProcMeansOptions} page for details on
the available options.}

\item{weight}{An optional weight parameter.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{options}{A vector of option keywords.}
}
\value{
The requested summary statistics.
}
\description{
The \code{proc_means} function generates summary statistics
for selected variables on the input dataset.  The variables are identified
on the \code{var} parameter.  The statistics to perform are identified
on the \code{stats} parameter.  By default, results are displayed in
the viewer interactively, and returned as a dataset or a list of datasets.
}
\details{
Some details about the sample function.

All statistics remove missing values by default.  Should there be an
option to include them?
}
\section{Options}{

\itemize{
\item{\strong{CSS}: Corrected Sum of Squares.

}
\item{\strong{CV}: Coefficient of Variation.

}
}
}

\section{Using Weights}{

Need a section explaining how to use weights.
}

\section{Data Constraints}{

Explain limits of data for each stat option.  Number of non-missing
values, etc.
}

