% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\encoding{UTF-8}
\name{proc_means}
\alias{proc_means}
\title{Calculates Summary Statistics}
\usage{
proc_means(
  data,
  by = NULL,
  class = NULL,
  var = NULL,
  stats = c("n", "mean", "std", "min", "max"),
  view = TRUE,
  report_type = NULL,
  report_location = NULL,
  report_style = NULL,
  titles = NULL,
  options = NULL,
  ...
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{by}{An optional by group. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{var}{The variable(s) to calculate summary statistics for. You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{stats}{A quoted vector of summary statistics keywords.  Valid
keywords are: "css", "clm", "cv", "n", "lclm", "mean", "median", "mode",
"min", "max",
"nmiss", "nobs", "range", "std", "stderr", "sum", "uclm", "var",
"q1", "q3", "p1", "p5", "p10", "p20", "p25", "p30", "p40",
"p50", "p60", "p70", "p75", "p80", "p90",
"p95", "p99", "qrange". You may
pass unquoted variable names to this parameter using the \code{\link{v}}
function.}

\item{view}{Whether to display the report in the interactive viewer.  Valid
values are TRUE and FALSE.  Default is TRUE.}

\item{report_type}{The output type for any report output.  Valid values are
'HTML', 'TXT', 'PDF', 'RTF', 'DOCX'.  Multiple outputs can be
requested by passing a vector of output types.  If the report_location parameter
is specified, the outputs will be written to that location.  Otherwise,
they will be written to a temp directory.  The default value is NULL.}

\item{report_location}{A path to write any report output requested by
the print parameter. The path may be either a full path with file name,
or a directory name.  If the file name is not specified, the procedure
type will be used.  If no path is specified, files will be
written to a temp directory.  Default is NULL.}

\item{report_style}{A theme name or style object to use on the report output.
See the \code{\link[reporter]{create_style}} in the \strong{reporter}
package for additional information on styles.}

\item{titles}{A vector of one or more titles to use for the report output.}

\item{options}{An options object created with the \code{\link{opts}}
function. This parameter can accept a variety of options.}

\item{...}{One or more output requests. To request a dataset output,
use the desired dataset name as the parameter name, and pass in the
\code{\link{output}} function with the desired parameters.  If only one
output dataset is requested, the function will return a single data frame.
If more than one output dataset is requested, the function will return
a list of data frames.}
}
\value{
The requested summary statistics.
}
\description{
The \code{proc_means} function generates summary statistics
for selected variable on the input dataset.  The variables are identified
on the \code{var} parameter.  The statistics to perform are identified
on the \code{stats} parameter.  By default, results are returned as a list,
of tibbles.
}
\details{
Some details about the sample function.

All statistics remove missing values by default.  Should there be an
option to include them?
}
\section{Statistics Options}{

\itemize{
\item{\strong{CSS}: Corrected Sum of Squares.

}
\item{\strong{CV}: Coefficient of Variation.

}
}
}

\section{Using Weights}{

Need a section explaining how to use weights.
}

\section{Data Constraints}{

Explain limits of data for each stat option.  Number of non-missing
values, etc.
}

