% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_sort.R
\name{proc_sort}
\alias{proc_sort}
\title{Sorts a data.frame or tibble}
\usage{
proc_sort(data, var = NULL, by = NULL, nodupkey = FALSE)
}
\arguments{
\item{data}{The input data to sort.}

\item{var}{The variables on the output data to keep.}

\item{by}{The variables to sort by.}

\item{nodupkey}{Whether to remove duplicates.}
}
\value{
The sorted dataset.  If a data frame was input, a
data frame will be output.  If a tibble was input, a tibble will
be output.
}
\description{
Here is a description
}
\details{
Here is some details
}
