% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_transpose.R
\encoding{UTF-8}
\name{proc_transpose}
\alias{proc_transpose}
\title{Transposes a Data Frame}
\usage{
proc_transpose(
  data,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL,
  name = "NAME",
  namelabel = NULL,
  prefix = NULL,
  delimiter = ".",
  suffix = NULL
)
}
\arguments{
\item{data}{The input data frame for which calculate summary statistics.}

\item{var}{The variable or variables to transpose.}

\item{id}{The variable or variables to use for the transposed column names.}

\item{idlabel}{The variable to use for the transposed column labels.}

\item{copy}{A vector of variables to retain in the output data
without transposition.}

\item{name}{Specifies the name of the variable to be used for the
var values.}

\item{namelabel}{The label to use for the name variable.}

\item{prefix}{Contains a prefix to be used in the construction of
column names.}

\item{delimiter}{Specifies a delimiter to be used in the construction
of column names.}

\item{suffix}{Contains a suffix to be used in the construction of
column names.}
}
\value{
A data frames that contains the transposed data.
}
\description{
Here is a sample function.
}
\details{
Some details about the sample function.
}
