% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_transpose.R
\encoding{UTF-8}
\name{proc_transpose}
\alias{proc_transpose}
\title{Transposes a Data Frame}
\usage{
proc_transpose(
  data,
  by = NULL,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL,
  name = "NAME",
  namelabel = NULL,
  prefix = NULL,
  delimiter = ".",
  suffix = NULL,
  where = NULL,
  log = TRUE
)
}
\arguments{
\item{data}{The input data to transpose.}

\item{by}{An optional by group.  Parameter accepts a vector of one or more
quoted variable names. If the by group is requested, the data will be subset
by that variable and the transpose function will
transpose each group and stack them together in a single table.}

\item{var}{The variable or variables to transpose.  Parameter accepts a vector
of quoted variable names.}

\item{id}{The variable or variables to use for the transposed column names.}

\item{idlabel}{The variable to use for the transposed column labels.}

\item{copy}{A vector of variables to retain in the output data
without transposition.}

\item{name}{Specifies the name of the variable to be used for the
var values.}

\item{namelabel}{The label to use for the name variable.}

\item{prefix}{Contains a prefix to be used in the construction of
column names.}

\item{delimiter}{Specifies a delimiter to be used in the construction
of column names.}

\item{suffix}{Contains a suffix to be used in the construction of
column names.}

\item{where}{An expression to filter the rows after the transform
is complete.  Use the \code{\link{expression}} function to define
the where clause.}

\item{log}{Whether or not to log the procedure.  Default is TRUE.
This parameter is used internally.}
}
\value{
A data frame that contains the transposed data. If a data frame
is input, a data frame will be output.  If a tibble is input, a tibble
will be output.
}
\description{
A function to pivot or transpose a data frame. In the default
usage, the variables identified by the parameter \code{var} are transposed
to become rows. The variable values in the parameter \code{id} become
the new column names.  The function has several more parameters to control
how variables are named in the transposed data set.
}
\details{
The \code{proc_tranpose} function takes an input data frame or tibble
and transposes the columns and rows.

Talk about vars

Talk about ID

Talk about by groups

Talk about copy

Talk about column naming

Talk about where clause
}
\examples{
# Prepare data
dat <- data.frame(CAT = rownames(USPersonalExpenditure),
                  USPersonalExpenditure, stringsAsFactors = FALSE,
                  row.names = NULL)[1:4, ]

# View data
dat
#                   CAT X1940 X1945 X1950 X1955 X1960
# 1    Food and Tobacco 22.20 44.50 59.60  73.2  86.8
# 2 Household Operation 10.50 15.50 29.00  36.5  46.2
# 3  Medical and Health  3.53  5.76  9.71  14.0  21.1
# 4       Personal Care  1.04  1.98  2.45   3.4   5.4

# Transpose Data
tdat <-  proc_transpose(dat, id = "CAT", name = "Year")

# View results
tdat
#   Year Food and Tobacco Household Operation Medical and Health Personal Care
# 1 X1940             22.2                10.5               3.53          1.04
# 2 X1945             44.5                15.5               5.76          1.98
# 3 X1950             59.6                29.0               9.71          2.45
# 4 X1955             73.2                36.5              14.00          3.40
# 5 X1960             86.8                46.2              21.10          5.40
}
