% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_transpose.R
\encoding{UTF-8}
\name{proc_transpose}
\alias{proc_transpose}
\title{Transposes a Data Frame}
\usage{
proc_transpose(
  data,
  by = NULL,
  var = NULL,
  id = NULL,
  idlabel = NULL,
  copy = NULL,
  name = "NAME",
  namelabel = NULL,
  prefix = NULL,
  delimiter = ".",
  suffix = NULL,
  where = NULL,
  log = TRUE
)
}
\arguments{
\item{data}{The input data frame for which will be transposed.}

\item{by}{An optional by group.  Parameter accepts a vector of one or more
quoted variable names. If the by group is requested, the data will be subset
by that variable and the transpose function will
transpose each group and stack them together in a single table.}

\item{var}{The variable or variables to transpose.  Parameter accepts a vector
of quoted variable names.}

\item{id}{The variable or variables to use for the transposed column names.}

\item{idlabel}{The variable to use for the transposed column labels.}

\item{copy}{A vector of variables to retain in the output data
without transposition.}

\item{name}{Specifies the name of the variable to be used for the
var values.}

\item{namelabel}{The label to use for the name variable.}

\item{prefix}{Contains a prefix to be used in the construction of
column names.}

\item{delimiter}{Specifies a delimiter to be used in the construction
of column names.}

\item{suffix}{Contains a suffix to be used in the construction of
column names.}

\item{where}{An expression to filter the rows after the transform
is complete.  Use the \code{\link{expression}} function to define
the where clause.}

\item{log}{Whether or not to log the procedure.  Default is TRUE.
This parameter is used internally.}
}
\value{
A data frame that contains the transposed data. If a data frame
is input, a data frame will be output.  If a tibble is input, a tibble
will be output.
}
\description{
A function to pivot or transpose a data frame. In the default
usage, the variables identified by the parameter \code{var} are transposed
to become rows. The variable values in the parameter \code{id} become
the new column names.  The function has several more parameters to control
how variables are named in the transposed data set.
}
\details{
The \code{proc_tranpose} function takes an input data frame or tibble
and transposes the columns and rows.
}
\examples{
# Create data
dt <- data.frame(names = rownames(mtcars), mtcars, stringsAsFactors = FALSE)[1:5,]

# Transpose data
tdt <- proc_transpose(dt, var = c("mpg", "cyl", "disp"),
                          id = "names", name = "Variable")

# View transposed data
tdt
#   Variable Mazda RX4 Mazda RX4 Wag Datsun 710 Hornet 4 Drive Hornet Sportabout
# 1      mpg        21            21       22.8           21.4              18.7
# 2      cyl         6             6        4.0            6.0               8.0
# 3     disp       160           160      108.0          258.0             360.0
}
