% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_up}
\alias{round_up}
\title{Simulates SAS® Rounding}
\usage{
round_up(x)
}
\arguments{
\item{x}{A vector of values to round.}
}
\value{
The rounded data vector.
}
\description{
A function to simulate the way SAS® performs rounding.
}
