% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestplots.R
\name{ttestplot}
\alias{ttestplot}
\title{Request T-Test Plots}
\usage{
ttestplot(type = "default", panel = TRUE, showh0 = FALSE)
}
\arguments{
\item{type}{The type(s) of plot to create. Multiple types should be passed
as a vector of strings.  Valid values are "agreement", "boxplot", "histogram",
"interval", "profiles", "qqplot", "summary".  The default value is
a vector with "summary" and "qqplot".}

\item{panel}{Whether or not to display the summary plot combined into
in a single panel.  Default is TRUE.  A value of FALSE will create
individual plots instead.  This parameter is equivalent to the
"unpack" keyword in SAS.}

\item{showh0}{Whether or not to show the h0 line on a single-variable T-test.
Default is FALSE.}
}
\description{
A function to request T-test plots on a call to
\code{\link{proc_ttest}}. The function allows you to specify the type of
regression plots to produce. It produces a combined diagnostics panel
by default.  You may also specify individual plots to create by setting
the "panel" parameter to FALSE, and passing a vector of plot names
on the "type" parameter.
}
\details{
Any requested
plots will be displayed on interactive reports only.
Plots are created as jpeg files, and stored in a temp directory.  Those
temporary files are then referenced by the interactive report to display
the graphic.

If desired, you may
output the report objects and pass to \code{\link{proc_print}}. To do this,
set \code{output = report} on the call to \code{\link{proc_freq}}, and pass
the entire list to \code{\link{proc_print}}.
}
\section{Plots}{

The \code{plots} parameter allows you to request several types of T-Test
plots. Below are the types of plots that are supported.  The list shows
the plot type keyword needed to request the plot, and a brief description:
\itemize{
\item{\strong{agreement}:  An agreement plot for the input data.  Only
available for paired comparisons.
}
\item{\strong{boxplot}: Displays a box and whisker plot for group comparisons,
including confidence intervals (if appropriate).
}
\item{\strong{histogram}: A histogram with normal curve and kernel density
overlays.
}
\item{\strong{interval}: Visualizes the confidence intervals for means.
}
\item{\strong{profiles}: A line plot mapping responses
between analysis variables. Available only for paired analysis.
}
\item{\strong{qqplot}: A quantile-quantile plot used to assess the assumption
of normality.
}
\item{\strong{summary}: Combines the histogram and boxplot charts
onto the same panel.
}
}
The above plots may be requested as a vector of keywords to the \code{plots}
parameter on \code{\link{proc_ttest}}, or as vector of values
to the \code{type} parameter on \code{ttestplots}.

Note that, when passed as a vector of keywords, only the requested plots
will be produced.  That is to say, the "only" keyword in SAS is implied
at all times for \code{\link{proc_ttest}}.
}

\examples{
library(procs)
}
