% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestplots.R
\name{ttestplot}
\alias{ttestplot}
\title{Request T-Test Plots}
\usage{
ttestplot(type = c("summary", "qqplot"), panel = TRUE, showh0 = FALSE)
}
\arguments{
\item{type}{The type(s) of plot to create. Multiple types should be passed
as a vector of strings.  Valid values are "agreement", "boxplot", "histogram",
"interval", "profiles", "qqplot", "summary".  The default value is
a vector with "summary" and "qqplot".  The "summary"
keyword produces a single combined chart with 2 different plots: a histogram
and a box plot.}

\item{panel}{Whether or not to display the summary plot combined into
in a single panel.  Default is TRUE.  A value of FALSE will create
individual plots instead.  This parameter is equivalent to the
"unpack" keyword in SAS.}

\item{showh0}{Whether or not to show the h0 line on a single-variable T-test.
Default is FALSE.}
}
\description{
A function to request T-test plots on a call to
\code{\link{proc_ttest}}. The function allows you to specify the type of
regression plots to produce. It produces a combined diagnostics panel
by default.  You may also specify individual plots to create by setting
the "panel" parameter to FALSE, and passing a vector of plot names
on the "type" parameter.
}
\details{
Any requested
plots will be displayed on interactive reports only.
Plots are created as jpeg files, and stored in a temp directory.  Those
temporary files are then referenced by the interactive report to display
the graphic.

If desired, you may
output the report objects and pass to \code{\link{proc_print}}. To do this,
set \code{output = report} on the call to \code{\link{proc_freq}}, and pass
the entire list to \code{\link{proc_print}}.
}
\section{Plots}{

The \code{plots} parameter allows you to request several types of regression
plots. Below are the types of plots that are supported.  The list shows
the plot type keyword needed to request the plot, and a brief description:
\itemize{
\item{\strong{diagnostics}:  A fit diagnostics panel that contains 8 different
types of plots and a table of statistics.
}
\item{\strong{residuals}: Produces a panel of residual plots against
each independent variable in the model.
}
\item{\strong{fitplot}:  Produces a scatter plot of the dependent variable
against the regressor, including the fitted line and confidence/prediction bands.
This is only available for models with a single regressor.
}
\item{\strong{qqplot}: Normal Quantile-Quantile (Q-Q) plot of residuals.
}
\item{\strong{rfplot}: Residual-Fit (RF) spread plot.
}
\item{\strong{residualbypredicted}: Residuals vs. Predicted values.
}
\item{\strong{rstudentbypredicted}: Externally Studentized Residuals (RStudent) vs. Predicted values.
}
\item{\strong{rstudentbyleverage}: Externally Studentized Residuals vs. Leverage.
}
\item{\strong{cooksd}: Cook’s D statistic vs. Observation number.
}
\item{\strong{residualhistogram}: Histogram of residuals,
with a normal and kernel curve overlay.
}
\item{\strong{observedbypredicted}: Dependent variable (Observed) vs. Predicted values.
}
\item{\strong{residualboxplot}: A boxplot of residuals.
}
}
The above plots may be requested in different ways: as a vector of keywords,
or as a call to the \code{\link{regplot}} function.  The keyword approach will
produce plots with default parameters. A call to \code{\link{regplot}} will
give you control over some parameters to the charts.  See the \code{\link{regplot}}
function for further details.
}

\section{Statistics}{

\itemize{
\item{\strong{adjrsq}: Adjusted R-square.
}
\item{\strong{aic}: Akaike's information criterion.
}
\item{\strong{coeffvar}: Coefficient of variation.
}
\item{\strong{depmean}: Mean of dependent.
}
\item{\strong{default}: A set of default statistics.
}
\item{\strong{edf}: Error degrees of freedom.
}
\item{\strong{mse}: Mean squared error.
}
\item{\strong{nobs}: Number of observations used.
}
\item{\strong{nparm}: Number of parameters in the model (including the intercept).
}
\item{\strong{rsquare}: The R-square statistic.
}
\item{\strong{sse}: Error sum of squares.
}
}
}

\examples{
library(procs)
}
